Download charts to local fileystem
To download charts from a repo, we have to interact a little with
Helm's support packages.

In particular, we need to resolve a repo URL, a chart name, and a
chart version, to a URL for the chart tarball; and, we need to do some
preliminaries to work with Helm's fetching infrastructure
(k8s.io/helm/pkg/getter).

After that, it remains only to GET the tarball and save it locally.